feat: module optimization
